user profile, admin get uesr profiles, deleter user by user and admin

+++ src/users/users.controller.ts
@@ -6,37 +6,61 @@ import {
   Patch,
   Param,
   Delete,
+  UseGuards,
+  ParseIntPipe,
+  HttpCode,
+  HttpStatus,
 } from '@nestjs/common';
 import { UsersService } from './users.service';
-import { CreateUserDto } from './dto/create-user.dto';
-import { UpdateUserDto } from './dto/update-user.dto';
+import { AuthGuard } from 'src/auth/guards/auth.guard';
+import { UpdateProfileDto } from './dto/update-user.dto';
+import { currentUser, Roles } from 'src/auth/decorators/auth.decrators';
+import { JwtPayload } from 'src/utils/types';
+import { AcceptFormData } from 'src/common/decorators/accept-form-data.decorator';
+import { userRole } from 'src/utils/enum';
 
-@Controller('users')
+@Controller('/api/user/')
 export class UsersController {
   constructor(private readonly usersService: UsersService) {}
 
-  @Post()
-  create(@Body() createUserDto: CreateUserDto) {
-    return this.usersService.create(createUserDto);
-  }
-
+  // User profile
   @Get()
-  findAll() {
-    return this.usersService.findAll();
+  @UseGuards(AuthGuard)
+  getProfileData(@currentUser() payload: JwtPayload) {
+    return this.usersService.getUserDetails(payload.id);
   }
 
+  // public profile
   @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.usersService.findOne(+id);
+  getPublicProfileData(@Param('id', ParseIntPipe) id: number) {
+    return this.usersService.getUserDetails(id);
   }
 
-  @Patch(':id')
-  update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
-    return this.usersService.update(+id, updateUserDto);
+  @Patch()
+  @AcceptFormData()
+  @UseGuards(AuthGuard)
+  updateProfile(
+    @Body() body: UpdateProfileDto,
+    @currentUser() payload: JwtPayload,
+  ) {
+    return this.usersService.updateProfile(body, payload);
   }
 
-  @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.usersService.remove(+id);
+  @Delete()
+  @UseGuards(AuthGuard)
+  @HttpCode(HttpStatus.NO_CONTENT)
+  deleteAccount(@currentUser() payload: JwtPayload) {
+    return this.usersService.remove(payload);
+  }
+  // Delete by admin
+  @Delete('/admin/:id')
+  @Roles(userRole.ADMIN)
+  @UseGuards(AuthGuard)
+  @HttpCode(HttpStatus.NO_CONTENT)
+  deleteUserAccount(
+    @Param('id', ParseIntPipe) id: number,
+    @currentUser() payload: JwtPayload,
+  ) {
+    return this.usersService.remove({ id, role: payload.role });
   }
 }
